Tweak dependabot config to bump dependencies in all dirs (#92)
With the previous config, we'd get several PRs for the same dependency bump. For example, for Python, two PRs were created for a dependency bump in the event-handler's requirements.txt file.

The intent for this new config is to bump each depdencency in each service with one PR. We can do this because no service requires a specific version of depdencency that's different from another service.

The man motivation for this PR was the `pip` package ecosystem, but the same applies for Docker as well.